Still working on verifying USM code is as expected. Seeing very small PWM signal but no motor movement at the current settings.

+++ mri_arm/motor_config.h
@@ -5,7 +5,7 @@
 typedef struct { /* __MOTOR_CONFIG_H */
     // Base Info
     int dir;                // Motor direction
-    int max_speed;          // max motor speed
+    float max_speed;          // max motor speed in rad/s
 
     // Encoder Into
     float qdec_cpr;         // motor encoder count per rotation
